完善任务系统

+++ game/src/util/BaseUtil.java
@@ -69,7 +69,7 @@ public class BaseUtil {
 	 * @param seconds
 	 * @return
 	 */
-	public static String secondsToDateStr( long seconds ){
+	public static String secondsToDateStr( int seconds ){
 		SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");//可以方便地修改日期格式
 		return dateFormat.format( new Date( seconds * 1000l ) );
